# docs/openapi.json
{
  "components": {
    "schemas": {
      "AnswerResponse": {
        "description": "A grounded answer. Citation fields are exactly the Ticket #10 contract.",
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationResponse"
            },
            "minItems": 1,
            "title": "Citations",
            "type": "array"
          },
          "refused": {
            "const": false,
            "title": "Refused",
            "type": "boolean"
          },
          "trace_id": {
            "format": "uuid",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "answer",
          "citations",
          "refused",
          "trace_id"
        ],
        "title": "AnswerResponse",
        "type": "object"
      },
...
      "AskRequest": {
        "additionalProperties": false,
        "properties": {
          "question": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Persist this exchange in the caller's own session.",
            "title": "Session Id"
          },
          "stream": {
            "default": false,
            "description": "Return grounded SSE when true.",
            "title": "Stream",
            "type": "boolean"
          }
        },
        "required": [
          "question"
        ],
        "title": "AskRequest",
        "type": "object"
      },
      "AskTokenData": {
        "properties": {
          "delta": {
            "title": "Delta",
            "type": "string"
          },
          "trace_id": {
            "format": "uuid",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "delta",
          "trace_id"
        ],
        "title": "AskTokenData",
        "type": "object"
      },
...
      "CitationResponse": {
        "description": "AC-2.5: exactly seven fields; score is sigmoid(BGE cross-encoder logit).",
        "properties": {
          "chunk_id": {
            "format": "uuid",
            "title": "Chunk Id",
            "type": "string"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "document_name": {
            "title": "Document Name",
            "type": "string"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "relevance_score": {
            "maximum": 1,
            "minimum": 0,
            "title": "Relevance Score",
            "type": "number"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          },
          "text_snippet": {
            "title": "Text Snippet",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "document_name",
          "section",
          "page",
          "chunk_id",
          "relevance_score",
          "text_snippet"
        ],
        "title": "CitationResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "pattern": "^[A-Z_]+$",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "description": "Safe human-readable explanation; never raw server details.",
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "RefusalResponse": {
        "description": "Successful safe refusal, returned with HTTP 200 rather than a server failure.",
        "properties": {
          "answer": {
            "const": "Not enough information in the corpus",
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationResponse"
            },
            "maxItems": 0,
            "title": "Citations",
            "type": "array"
          },
          "refused": {
            "const": true,
            "title": "Refused",
            "type": "boolean"
          },
          "trace_id": {
            "format": "uuid",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "answer",
          "citations",
          "refused",
          "trace_id"
        ],
        "title": "RefusalResponse",
        "type": "object"
      },
...
      "RetrieveRequest": {
        "additionalProperties": false,
        "properties": {
          "query": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "RetrieveRequest",
        "type": "object"
      },
      "RetrieveResponse": {
        "properties": {
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationResponse"
            },
            "title": "Citations",
            "type": "array"
          },
          "trace_id": {
            "format": "uuid",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "citations"
        ],
        "title": "RetrieveResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "JWT access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Domain Copilot HTTP API. Bearer JWT identity and stored roles are checked server-side. See docs/API-CONTRACTS.md for ownership and SSE semantics.",
    "title": "Domain Copilot",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/ask": {
      "post": {
        "description": "All roles may ask. Optional session_id must belong to the caller, including admins.\n\nstream=false returns JSON; stream=true returns SSE after grounding completes.\nA refusal is HTTP 200 with refused=true, the exact refusal text and empty citations.\nThis informational path never drafts, approves or finalizes a clinical note.\n\nRoles: analyst, reviewer, admin. Session resources are owner-only for every role, including admin. Identity and current role come from the validated JWT and stored user.",
        "operationId": "ask_api_v1_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "False": "#/components/schemas/AnswerResponse",
                      "True": "#/components/schemas/RefusalResponse"
                    },
                    "propertyName": "refused"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/AnswerResponse"
                    },
                    {
                      "$ref": "#/components/schemas/RefusalResponse"
                    }
                  ],
                  "title": "Response Ask Api V1 Ask Post"
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "DOMAIN_ERROR, APPLICATION_ERROR or HTTP_ERROR: invalid operation."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "NOT_AUTHENTICATED. WWW-Authenticate: Bearer.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "const": "Bearer",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PERMISSION_DENIED or RESOURCE_FORBIDDEN."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RESOURCE_NOT_FOUND or HTTP_ERROR."
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "HTTP_ERROR: method not allowed."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_STATE_TRANSITION: conflict with persisted state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "UPLOAD_TOO_LARGE."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "VALIDATION_ERROR or INVARIANT_VIOLATION: invalid input."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR or CONFIGURATION_ERROR. Static, sanitized detail."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JOB_STORE_UNAVAILABLE, KNOWLEDGE_UNAVAILABLE, APPROVAL_STORE_UNAVAILABLE or HISTORY_STORE_UNAVAILABLE. Static, sanitized detail."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Ask",
        "tags": [
          "knowledge"
        ],
        "x-ownership": "Session resources are owner-only for every role, including admin.",
        "x-roles": [
          "analyst",
          "reviewer",
          "admin"
        ],
        "x-sse-events": {
          "refusal": {
            "$ref": "#/components/schemas/RefusalResponse"
          },
          "stream_completed": {
            "$ref": "#/components/schemas/AnswerResponse"
          },
          "token": {
            "$ref": "#/components/schemas/AskTokenData"
          }
        }
      }
    },
...
    "/api/v1/retrieve": {
      "post": {
        "description": "Roles: analyst, reviewer, admin. Authenticated caller; no owner-selected resource. Identity and current role come from the validated JWT and stored user.",
        "operationId": "retrieve_api_v1_retrieve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrieveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrieveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "DOMAIN_ERROR, APPLICATION_ERROR or HTTP_ERROR: invalid operation."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "NOT_AUTHENTICATED. WWW-Authenticate: Bearer.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "const": "Bearer",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PERMISSION_DENIED or RESOURCE_FORBIDDEN."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RESOURCE_NOT_FOUND or HTTP_ERROR."
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "HTTP_ERROR: method not allowed."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_STATE_TRANSITION: conflict with persisted state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "UPLOAD_TOO_LARGE."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "VALIDATION_ERROR or INVARIANT_VIOLATION: invalid input."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR or CONFIGURATION_ERROR. Static, sanitized detail."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JOB_STORE_UNAVAILABLE, KNOWLEDGE_UNAVAILABLE, APPROVAL_STORE_UNAVAILABLE or HISTORY_STORE_UNAVAILABLE. Static, sanitized detail."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Retrieve",
        "tags": [
          "knowledge"
        ],
        "x-ownership": "Authenticated caller; no owner-selected resource.",
        "x-roles": [
          "analyst",
          "reviewer",
          "admin"
        ]
      }
    },
...
  }
}